refactor(container): mv ThemeWrapper -> layout/ThemePalette
