+++ src/components/MarkDownRender/MathVersion.js
@@ -21,7 +21,7 @@ import mentionsPlugin from 'remarkable-mentions'
 import latexPlugin from 'remarkable-katex'
 import Prism from 'mastani-codehighlight'
 
-import MarkDownStyle from '@/containers/ThemeWrapper/MarkDownStyle'
+import MarkDownStyle from '@/containers/layout/ThemePalette/MarkDownStyle'
 import { MENTION_USER_ADDR } from '@/config'
 import { buildLog } from '@/utils'
 import { PreviewerContainer } from './styles'
