+++ src/components/MarkDownRender/index.js
@@ -15,7 +15,7 @@ import Prism from 'mastani-codehighlight'
 import { MENTION_USER_ADDR } from '@/config'
 import { buildLog } from '@/utils'
 
-import MarkDownStyle from '@/containers/ThemeWrapper/MarkDownStyle'
+import MarkDownStyle from '@/containers/layout/ThemePalette/MarkDownStyle'
 import { PreviewerContainer } from './styles'
 
 /* eslint-disable-next-line */
