+++ src/containers/layout/ThemePalette/CodeSyxHighlight.js
