+++ src/containers/layout/ThemePalette/MarkDownStyle.js
