+++ src/containers/layout/ThemePalette/RichEditorStyle.js
