+++ src/pages/_error.js
@@ -4,7 +4,7 @@ import { Provider } from 'mobx-react'
 import * as Sentry from '@sentry/node'
 
 import AnalysisService from '@/services/Analysis'
-import ThemeWrapper from '@/containers/ThemeWrapper'
+import ThemePalette from '@/containers/layout/ThemePalette'
 import ErrorPage from '@/components/ErrorPage'
 
 import { useStore } from '@/stores/init'
@@ -26,11 +26,11 @@ const Error = (props) => {
   return (
     <Provider store={store}>
       <AnalysisService>
-        <ThemeWrapper>
+        <ThemePalette>
           <Wrapper>
             <ErrorPage errorCode={statusCode} />
           </Wrapper>
-        </ThemeWrapper>
+        </ThemePalette>
       </AnalysisService>
     </Provider>
   )
